+++ database/migrations/2022_10_31_141446_add_table_app_links.php
@@ -1,0 +1,34 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+return new class extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::create('apps_links', function (Blueprint $table) {
+            $table->integer('id')->primaryKey();
+            $table->integer('app_id')->nullable()->index()->default(0);
+            $table->string('name');
+            $table->string('url')->nullable();
+            $table->string('icon')->nullable();
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        //
+    }
+};