+++ database/migrations/2022_10_31_143033_add_columns_to_app_links.php
@@ -1,0 +1,34 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+use Carbon\Carbon;
+
+return new class extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('apps_links', function (Blueprint $table) {
+            $table->dateTime('created_at')->default(Carbon::now()->toDateTimeString());
+            $table->dateTime('updated_at')->default(Carbon::now()->toDateTimeString());
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('app_links', function (Blueprint $table) {
+            //
+        });
+    }
+};