+++ resources/views/home/index.blade.php
@@ -1,11 +1,21 @@
 @extends('layouts.app-master')
 
 @section('content')
-    <div class="bg-light p-5 rounded">
+    <div class=" p-5 rounded">
         @auth
-        <h1>Dashboard</h1>
-        @include('home.partials.search')
-      
+        
+
+        <div class="px-4 py-5 my-5 text-center">
+            {{-- <img class="d-block mx-auto mb-4" src="../assets/brand/bootstrap-logo.svg" alt="" width="72" height="57"> --}}
+            <h1 class="display-5 fw-bold">Shortcuts</h1>
+            <div class="col-lg-6 mx-auto">
+              <p class="lead mb-4"></p>
+              <div class="d-grid gap-2 d-sm-flex justify-content-sm-center">
+                <a href="{{route('home.results')}}" type="button" class="btn btn-primary btn-lg px-4 gap-3">Search documentation <i class="fa fa-search"></i></a>
+                <a href="{{route('documents.add_show')}}" type="button" class="btn btn-outline-secondary btn-lg px-4">Add documentation <i class="fa fa-plus"></i></a>
+              </div>
+            </div>
+          </div>
         @endauth
 
         @guest